Uploading Classification code
Having an issue with the classification report, may have an issue with the prediction part. Accuracy is also extremely low, need to make adjustments. The loop for different KN neighbor values is not working correctly. Everything else is working fine, graph is also correct and what I wanted. 

Just need to figure out the accuracy, classification, and K loop portion.